SFW-41 test cases refactoring
